backend/requirements.txt
async-timeout==5.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
scapy==2.6.1
six==1.16.0
spotipy==2.23.0
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
